broaden value range for glad and tcl boolean values, create standard and custom feature filters. Add Git SHA1 to Fat Jar filename
